+++ README.md
@@ -2,13 +2,14 @@
 
 # Stitch
 
-# Quickstart
+## Quickstart
 
 Run `cargo run --release --bin=compress data/cogsci/nuts-bolts.json --max-arity=2 --iterations=3`
 
 In around 10 seconds this should produce an output like
 
 ```
+
 =======Compression Summary=======
 Found 3 inventions
 Cost Improvement: (3.92x better) 1919558 -> 489600
@@ -20,6 +21,7 @@ Wrote to "out/out.json"
 ```
 
 Brief guide to reading this:
+
 - `inv0` is the autogenerated name of the invention
 - `(1.78x wrt orig)` means the resulting compressed programs using `inv0` were 1.78x smaller than the original programs
 - `utility: 836528` this is a measure of how many much smaller the program got after rewriting it in terms of the new primitives (divide by 100 to get appx how  many fewer leaf nodes are in the rewritten program)
@@ -31,10 +33,31 @@ Brief guide to reading this:
 To see a full list of command line options run `cargo run --release --bin=compress -- --help`
 
 ## Disabling optimizations
+
 `cargo run --release --bin=compress data/cogsci/nuts-bolts.json -i3 --no-opt-free-vars --no-opt-upper-bound --no-opt-single-use --no-opt-force-multiuse`
+
+## Benchmarking
+
+* `cargo bench` runs the benchmarks. Running it twice in a row (e.g. from different branches) will do a comparison
+
+Comparing your feature before merging it:
+
+```bash
+git checkout main
+cargo bench --bench=compress_bench -- --save-baseline=main
+git checkout feature
+cargo bench --bench=compress_bench -- --save-baseline=feature
+cargo bench -- --load-baseline=feature --baseline=master
+```
+
+Details:
+
+- `--save-baseline=main` saves a named baseline (comparing against a past version of it if it exists, then overwriting it)
+- `--load-baseline=feature` means *don't run any benchmarks* just load the file as if it's a result that you just produced
+- `--baseline=master` overrides which benchmark we're going to compare against
+- `--bench=compress_bench` avoids the "unrecognized option" error detailed [here](https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options)
 
 ## Flamegraph
 
+install if you havent: `cargo install flamegraph`
 `cargo build --release && sudo cargo flamegraph --root --output=out/flamegraph.svg --bin=compress data/cogsci/nuts-bolts.json`
-
-
